Add files via upload
o	IoT-based project. C++ language has been used to build this project. Arduino uno has been used as the main microcontroller. Other components are a 500-watt DC to AC inverter, relay, seven segment display to see battery level, LEDs and push button.

o	The main aim of this work is to use solar energy in efficient way. Typically, in our country most of the solar system is used to charge the battery after that energy is supplied from the battery. It is not an efficient way because the battery is fully charged earlier and the sunlight energy wasted during the day.

o	When the battery is fully charged the system cuts the main grid electricity and generates AC energy from the solar panel and supply to the electrical component in the house and when the sun goes out and the battery is drained to a certain level (level can set using push button) the system automatically turn on the main grid.